+++ frontend/lost/src/containers/Annotation/SIA/lost-sia/CHANGELOG.md
@@ -5,9 +5,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
-## [unreleased] - 
+## [0.9.0] - 2021-12-23
 ### Added
 - Delete annotation in creation mode when hitting *Escape*-Key
+- Show img description in ImgBar if available
+- Max canvas size mode. Where canvas takes the maximum container size and is not 
+  image oriented as before. Add prop *maxCanvas={true}* to canvas in order to enable.
 ### Changed
 - Be able to deal with mixed color possible labels -> where a part of labels 
   will have a specified color and the other part has no provided color